requirements.txt
python-dotenv
pandas
numpy
scikit-learn
matplotlib
seaborn
flask
requests
pymongo
setuptools
certifi
pymongo[srv]==3.12
PyYAML
dill
